fix: increase size of some integer types

+++ migrations/000_initial.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS users (
   password TEXT NOT NULL,
   enabled BOOLEAN NOT NULL,
   admin BOOLEAN NOT NULL,
-  "limit" INTEGER,
+  "limit" BIGINT,
   created_at DATETIME NOT NULL,
   created_by INTEGER
 );
@@ -15,10 +15,10 @@ CREATE TABLE IF NOT EXISTS uploads (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
   slug TEXT NOT NULL,
   filename TEXT NOT NULL,
-  size INTEGER NOT NULL,
+  size BIGINT NOT NULL,
   public BOOLEAN NOT NULL,
-  downloads INTEGER NOT NULL,
-  "limit" INTEGER,
+  downloads BIGINT NOT NULL,
+  "limit" BIGINT,
   expiry_date DATE,
   uploaded_by INTEGER NOT NULL REFERENCES users (id),
   uploaded_at DATETIME NOT NULL,
